Add Data.Scientific.Base
